+++ styx-flyte-client/src/main/java/com/spotify/styx/flyte/client/FlyteInputsUtils.java
@@ -149,21 +149,4 @@ public class FlyteInputsUtils {
     inputsMap.putAll(map2); // Then override with the triggeredParams
     return ImmutableMap.copyOf(inputsMap);
   }
-
-  // Test this filter by printing it and using it on a flytectl command
-  // example:  flytectl get executions -p PROJECT -d DOMAIN --filter.fieldSelector="execution.phase in (RUNNING),execution.started_at>2022-02-07T18:23:05,execution.started_at<2022-02-08T18:10:05" -o json
-  public static String getExecutionsListFilter(Instant timeNow) {
-
-    String minus3m = timeNow.minus(3, ChronoUnit.MINUTES)
-        .atZone(ZoneId.of("UTC"))
-        .toLocalDateTime()
-        .format(DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss"));
-
-    final String minus24h = timeNow.minus(24, ChronoUnit.HOURS)
-        .atZone(ZoneId.of("UTC"))
-        .toLocalDateTime()
-        .format(DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss"));
-
-    return String.format("execution.phase in (RUNNING),execution.started_at>%s,execution.started_at<%s", minus24h, minus3m);
-  }
 }
